Rework of Kafka writer
* Moved Kafka writer to its own module to have better dependency managment
 * Simplified and cleaned up Kafka writer
 * Added minimal assertions to tests
