+++ jmxtrans-output/pom.xml
@@ -31,6 +31,7 @@
 		<module>jmxtrans-output-jrobin</module>
 		<module>jmxtrans-output-log4j</module>
 		<module>jmxtrans-output-velocity</module>
+		<module>jmxtrans-output-kafka</module>
 	</modules>
 
 	<build>
