+++ jmxtrans/pom.xml
@@ -11,12 +11,10 @@
 	<artifactId>jmxtrans</artifactId>
 	<packaging>jar</packaging>
 	<name>JmxTrans</name>
-	<description>
-		JMX metrics exporter.
+	<description>JMX metrics exporter.
 
 		This module creates the packaging for JmxTrans. It does not contain any application code, but creates .deb, .rpm
-		or other packaging.
-	</description>
+		or other packaging.</description>
 
 	<properties>
 		<verify.mutationThreshold>0</verify.mutationThreshold>
@@ -82,6 +80,11 @@
 		<dependency>
 			<groupId>org.jmxtrans</groupId>
 			<artifactId>jmxtrans-output-jrobin</artifactId>
+			<scope>runtime</scope>
+		</dependency>
+		<dependency>
+			<groupId>org.jmxtrans</groupId>
+			<artifactId>jmxtrans-output-kafka</artifactId>
 			<scope>runtime</scope>
 		</dependency>
 		<dependency>
